Move tests to server folder
