+++ server/test/api.v1/questions.spec.js
@@ -1,6 +1,6 @@
 import chai, { expect } from 'chai';
 import chaiHttp from 'chai-http';
-import server from '../../app';
+import server from '../../../app';
 
 chai.use(chaiHttp);
